Change manual build dir to doc/manual-src/*/_build. Drop html install.
Sphinx HTML markups are beautiful, but it uses many asset files, such
as stylesheets, javascripts and images.  While it is no problem to
upload them in web site, but distributing and installing them in
individual user have some problems.  For example, Sphinx uses same
assets for generated sites, so if many applications uses Sphinx
generated docs and they are installed in the PC, it is huge waste of
the disk space. I'm also not comfortable to copy HTML trees to the
install directory using 'cp -r' command.  I seeked other format like
pdf and texi which Sphinx can generate but annoyingly they all convert
'--' into en-dash and there is no workaround to disable it. So I
decided to drop HTML manual from distribution and installation.  For
users who want to HTML version manual, see it online:
http://aria2.sourceforge.net/manual/en/html/