Make AX_CXX_COMPILE_STDCXX_11 test for -stdlib=libc++ via std::shared_ptr
The clang shipped with OSX XCode and clangs not build enabling libcpp,
will default to the libstdc++ headers and lib installed on the system.
In the OSX case, that libstdc++ is the one bundles with gcc-4.2, which
is far too old to provide all required C++11 types, such as
std::shared_ptr.
Hence, the C++11 check should try to compile a program with a C++11 type
and try -stdlib=libc++ if the default lib fails to compile said program.