Fix out-of-tree builds of libaria2api
Also it's not just the `html` target that needs libaria2api generated.

The fix is mighty ackward, as it is not really out-of-tree: libaria2api
still will be but into the srcdir. I settled for this because trying to
do a real include from the build dir is really a mess, either requireing
heavy pre-processing or a custom sphinx Include directive :p