Remove trac-wiki link
sf.net pulled the plug of hosted apps including trac-wiki.  The link
to the wiki has not worked for some time now.  We submit issue to
sf.net to migrate the wiki content to current sf.net standard wiki.
When the migration is done, we'll add the link to it in the doc.