Apply --retry-wait when retrying after 404
Previously, --retry-wait was not used when retrying after 404
(--max-file-not-found).  The retry was done without delay.  Also the
maximum retry count (--max-retries) was ignored when retrying with
404.

This commit makes --retry-wait apply for retrying with 404 as well.
It also combines --max-retries and --max-file-not-found.  When
retrying with 404, it is counted toward retry count.