Simplify cache write
The previous cache write routine was too complex.  I'm sure I can
rewrite it to more elegantly.  But the primary motivation of this
complex logic is for disk activity reduction on Windows 7, and I
observed it on my old IDE disk.  I checked it again recently, and
there is no difference between with and without this complex logic.
For this reason, it was removed.  Will revert this change if many of
users are not happy with this.