reliably identify zero-length files
Both zero-length files and files of unknown size have identical Range
specs. To identify zero-length, code relied on the presence of
content-length header or the absence of transfer-encoding. However,
these assumptions may not hold true as the length can also be specified
in content-range.

This change introduces an backward-compatible change to Range to
reliably distinguish between the two.